requirements.txt
fastai
dtreeviz
fastbook
pandas
scikit-learn
xgboost
seaborn
IPython
auto-sklearn
pytorch-tabnet
numpy
